Move self profile result parsing from `rustc-fake` to `collector`
